Add "terminal length 0" as fallback command to disable pager
Added an alternative command to disable pager in the Cisco ASA device terminal.

This change allows the script to continue functioning even if the 'no terminal pager' command fails for any reason. In such a case, the script will attempt the 'terminal length 0' command as an alternative.

This change helps to increase the robustness of the script and its ability to handle unexpected situations.